Updating to new Dialogue System
The Brackeys system was somewhat limited and complicated. It works, but if I wanted to change characters and stuff, I'd have to literallly make billions of copies of the two scripts. That wasn't gonna be fun.

I switched to the Blackthornprod system, which seems to be working better. I just need to make choices  a thing now.