Dialogue Management Complete
For now, of course. It's not very efficient for anything proper, but it works for a prototype. If I want to continue work on this, I'd have to probably rework a lot of things.

Nonetheless, there are some things I will need to fix. Most of it is making it so that I don't have to rely on hardcoding things, which means less script duplicates. Otherwise, I better get started on that story soon.